Refactor repeat logic with shared state
